+++ kitty/options/definition.py
@@ -1457,6 +1457,15 @@ by adding :code:`count-background` to the setting, for example: :code:`-1 count-
 Note that if you want confirmation when closing individual windows,
 you can map the :ac:`close_window_with_confirmation` action.
 ''')
+
+opt('window_drag_tolerance', '2', option_type='float', ctype='double', long_text='''
+Control dragging window borders to resize kitty windows. This is the tolerance in pts
+for the region around window borders where pressing the left mouse button
+will start the dragging of window borders. Use a large negative value such as -200 to disable
+dragging of borders. Note that because kitty uses layouts, dragging borders does not
+actually resize the window itself, but instead, the layout row/column/slot, which can result
+in multiple windows getting resized.
+''')
 egr()  # }}}
 
 
@@ -3846,6 +3855,21 @@ Remove all shortcut definitions up to this point. Useful, for instance, to
 remove the default shortcuts.
 '''
     )
+
+opt('map_timeout', '0.0', option_type='positive_float', long_text='''
+The default timeout (in seconds) for multi-key mappings and modal keyboard modes.
+If you press the first key(s) of a multi-key mapping and don't press the next
+key within this timeout, the mapping is cancelled and the mode is exited. A value
+of zero disables the timeout. This can be overridden for specific modes using the
+:code:`--timeout` option when creating a keyboard mode with :code:`--new-mode`.
+For example::
+
+    # 2 second timeout for all mappings
+    map_timeout 2.0
+
+    # This mode will have a 5 second timeout (overrides the global 2 second timeout)
+    map --new-mode resize --timeout 5.0 kitty_mod+r
+''')
 
 opt('+action_alias', 'launch_tab launch --type=tab --cwd=current',
     option_type='action_alias',