+++ kitty/options/parse.py
@@ -1120,6 +1120,9 @@ class Parser:
     def macos_window_resizable(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['macos_window_resizable'] = to_bool(val)
 
+    def map_timeout(self, val: str, ans: dict[str, typing.Any]) -> None:
+        ans['map_timeout'] = positive_float(val)
+
     def mark1_background(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['mark1_background'] = to_color(val)
 
@@ -1468,6 +1471,9 @@ class Parser:
 
     def window_border_width(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['window_border_width'] = window_border_width(val)
+
+    def window_drag_tolerance(self, val: str, ans: dict[str, typing.Any]) -> None:
+        ans['window_drag_tolerance'] = float(val)
 
     def window_logo_alpha(self, val: str, ans: dict[str, typing.Any]) -> None:
         ans['window_logo_alpha'] = unit_float(val)