+++ kitty/options/types.py
@@ -390,6 +390,7 @@ option_names = (
     'macos_traditional_fullscreen',
     'macos_window_resizable',
     'map',
+    'map_timeout',
     'mark1_background',
     'mark1_foreground',
     'mark2_background',
@@ -487,6 +488,7 @@ option_names = (
     'wheel_scroll_multiplier',
     'window_alert_on_bell',
     'window_border_width',
+    'window_drag_tolerance',
     'window_logo_alpha',
     'window_logo_path',
     'window_logo_position',
@@ -587,6 +589,7 @@ class Options:
     macos_titlebar_color: int = 0
     macos_traditional_fullscreen: bool = False
     macos_window_resizable: bool = True
+    map_timeout: float = 0
     mark1_background: Color = Color(152, 211, 203)
     mark1_foreground: Color = Color(0, 0, 0)
     mark2_background: Color = Color(242, 220, 211)
@@ -677,6 +680,7 @@ class Options:
     wheel_scroll_multiplier: float = 5.0
     window_alert_on_bell: bool = True
     window_border_width: tuple[float, str] = (0.5, 'pt')
+    window_drag_tolerance: float = 2.0
     window_logo_alpha: float = 0.5
     window_logo_path: str | None = None
     window_logo_position: choices_for_window_logo_position = 'bottom-right'